backend/fastapi-backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
pytest
pytest-asyncio
pymongo
python-dotenv
bson
